Fixed Bandwidth

+++ app/src/main/java/me/iologic/apps/dtn/FileServices.java
@@ -89,7 +89,7 @@ public class FileServices {
         return readData;
     }
 
-    public float getFileSize(){
+    public long getFileSize(){
         return readData.length;
     }
